Split state-related code into header and source
