+++ include/kettle/state/state.hpp
@@ -1,17 +1,10 @@
 #pragma once
 
-#include <algorithm>
-#include <cmath>
 #include <complex>
-#include <iomanip>
-#include <sstream>
-#include <stdexcept>
 #include <string>
-#include <utility>
 #include <vector>
 
 #include "kettle/common/mathtools.hpp"
-#include "kettle/state/bitstring_utils.hpp"
 #include "kettle/state/endian.hpp"
 #include "kettle/state/qubit_state_conversion.hpp"
 
@@ -28,50 +21,17 @@ public:
         The 0 state is the same in both the little and big endian representations, so it isn't needed
         in this constructor.
     */
-    explicit QuantumState(std::size_t n_qubits)
-        : n_qubits_ {n_qubits}
-        , n_states_ {ket::internal::pow_2_int(n_qubits)}
-        , coefficients_(n_states_, {0.0, 0.0})
-    {
-        check_at_least_one_qubit_();
-        coefficients_[0] = {1, 0};
-    }
+    explicit QuantumState(std::size_t n_qubits);
 
     explicit QuantumState(
         std::vector<std::complex<double>> coefficients,
         QuantumStateEndian input_endian = QuantumStateEndian::LITTLE
-    )
-        : n_qubits_ {0}  // can't properly set number of qubits before verifying coefficients
-        , n_states_ {coefficients.size()}
-        , coefficients_ {std::move(coefficients)}
-    {
-        check_power_of_2_with_at_least_one_qubit_();
-        check_normalization_of_coefficients_();
-
-        // the size of the coefficients vector is 2^{number of 0's in bit rep of the size}
-        n_qubits_ = static_cast<std::size_t>(std::countr_zero(coefficients_.size()));
-
-        // the user passed the vector of coefficients in big endian format, but the
-        // internal mapping of the indices to the states is in little endian format,
-        // so we need to perform a conversion
-        if (input_endian == QuantumStateEndian::BIG) {
-            perform_endian_flip_on_coefficients_();
-        }
-    }
+    );
 
     explicit QuantumState(
         const std::string& computational_state,
         QuantumStateEndian input_endian = QuantumStateEndian::LITTLE
-    )
-        : n_qubits_ {computational_state.size()}
-        , n_states_ {ket::internal::pow_2_int(computational_state.size())}
-        , coefficients_(n_states_, {0.0, 0.0})
-    {
-        impl_ket::check_bitstring_is_valid_nonmarginal_(computational_state);
-
-        const auto index = bitstring_to_state_index(computational_state, input_endian);
-        coefficients_[index] = {1.0, 0.0};
-    }
+    );
 
     constexpr auto operator[](std::size_t index) const noexcept -> const std::complex<double>&
     {
@@ -84,13 +44,13 @@ public:
     }
 
     [[nodiscard]]
-    constexpr auto at(std::size_t index) const -> const std::complex<double>&
+    auto at(std::size_t index) const -> const std::complex<double>&
     {
         check_index_(index);
         return coefficients_[index];
     }
 
-    constexpr auto at(std::size_t index) -> std::complex<double>&
+    auto at(std::size_t index) -> std::complex<double>&
     {
         check_index_(index);
         return coefficients_[index];
@@ -134,119 +94,27 @@ private:
     std::size_t n_states_;
     std::vector<std::complex<double>> coefficients_;
 
-    void check_power_of_2_with_at_least_one_qubit_() const
-    {
-        if (coefficients_.size() < 2) {
-            throw std::runtime_error {
-                "There must be at least 2 coefficients, representing the states for one qubit.\n"};
-        }
+    void check_power_of_2_with_at_least_one_qubit_() const;
 
-        if (!ket::internal::is_power_of_2(coefficients_.size())) {
-            auto err_msg = std::stringstream {};
-            err_msg << "The provided coefficients must have a size equal to a power of 2.\n";
-            err_msg << "Found size = " << coefficients_.size();
-            throw std::runtime_error {err_msg.str()};
-        }
-    }
+    void check_normalization_of_coefficients_() const;
 
-    void check_normalization_of_coefficients_() const
-    {
-        auto sum_of_squared_norms = double {0.0};
-        for (const auto& elem : coefficients_) {
-            sum_of_squared_norms += std::norm(elem);
-        }
+    void check_index_(std::size_t index) const;
 
-        const auto expected = 1.0;
-        const auto is_normalized = std::fabs(sum_of_squared_norms - expected) < ket::internal::NORMALIZATION_TOLERANCE;
+    void check_at_least_one_qubit_() const;
 
-        if (!is_normalized) {
-            auto err_msg = std::stringstream {};
-            err_msg << "The provided coefficients are not properly normalized.\n";
-            err_msg << "Found sum of squared norms : ";
-            err_msg << std::fixed << std::setprecision(14) << sum_of_squared_norms;
-            throw std::runtime_error {err_msg.str()};
-        }
-    }
-
-    constexpr void check_index_(std::size_t index) const
-    {
-        if (index >= n_states_) {
-            throw std::runtime_error {"Out-of-bounds access for the quantum state.\n"};
-        }
-    }
-
-    constexpr void check_at_least_one_qubit_() const
-    {
-        if (n_qubits_ == 0) {
-            throw std::runtime_error {"There must be at least 1 qubit in the QuantumState.\n"};
-        }
-    }
-
-    constexpr void perform_endian_flip_on_coefficients_() noexcept
-    {
-        for (std::size_t i {0}; i < n_states_; ++i) {
-            const auto i_flip = impl_ket::endian_flip_(i, n_qubits_);
-            if (i < i_flip) {
-                std::swap(coefficients_[i], coefficients_[i_flip]);
-            }
-        }
-    }
+    void perform_endian_flip_on_coefficients_() noexcept;
 };
 
-constexpr auto almost_eq(
+auto almost_eq(
     const QuantumState& left,
     const QuantumState& right,
     double tolerance_sq = ket::internal::COMPLEX_ALMOST_EQ_TOLERANCE_SQ
-) noexcept -> bool
-{
-    if (left.n_qubits() != right.n_qubits()) {
-        return false;
-    }
+) noexcept -> bool;
 
-    for (std::size_t i {0}; i < left.n_states(); ++i) {
-        if (!almost_eq(left[i], right[i], tolerance_sq)) {
-            return false;
-        }
-    }
+auto tensor_product(const QuantumState& left, const QuantumState& right) -> QuantumState;
 
-    return true;
-}
+auto inner_product(const QuantumState& bra_state, const QuantumState& ket_state) -> std::complex<double>;
 
-inline auto tensor_product(const QuantumState& left, const QuantumState& right) -> QuantumState
-{
-    const auto n_states = left.n_states() * right.n_states();
-    auto new_coefficients = std::vector<std::complex<double>> {};
-    new_coefficients.reserve(n_states);
-
-    for (std::size_t i_right {0}; i_right < right.n_states(); ++i_right) {
-        for (std::size_t i_left {0}; i_left < left.n_states(); ++i_left) {
-            new_coefficients.push_back(left[i_left] * right[i_right]);
-        }
-    }
-
-    return QuantumState {std::move(new_coefficients)};
-}
-
-inline auto inner_product(const QuantumState& bra_state, const QuantumState& ket_state) -> std::complex<double>
-{
-    if (bra_state.n_states() != ket_state.n_states()) {
-        throw std::runtime_error {"ERROR: cannot calculate inner product between two states of different sizes.\n"};
-    }
-
-    // calculate the inner product
-    auto inner_product = std::complex<double> {};
-    for (std::size_t i {0}; i < bra_state.n_states(); ++i) {
-        inner_product += (std::conj(bra_state[i]) * ket_state[i]);
-    }
-
-    return inner_product;
-}
-
-inline auto inner_product_norm_squared(const QuantumState& left, const QuantumState& right) -> double
-{
-    const auto inner_product_ = inner_product(left, right);
-
-    return std::norm(inner_product_);
-}
+auto inner_product_norm_squared(const QuantumState& left, const QuantumState& right) -> double;
 
 }  // namespace ket
