+++ source/kettle_internal/state/bitstring_utils.cpp
@@ -1,16 +1,15 @@
-#pragma once
-
 #include <algorithm>
 #include <cstddef>
 #include <stdexcept>
 #include <string>
 
 #include "kettle_internal/common/utils_internal.hpp"
+#include "kettle_internal/state/bitstring_utils.hpp"
 
-namespace impl_ket
+namespace ket::internal
 {
 
-constexpr auto endian_flip_(std::size_t value, std::size_t n_relevant_bits) -> std::size_t
+auto endian_flip_(std::size_t value, std::size_t n_relevant_bits) -> std::size_t
 {
     auto backward = std::size_t {0};
 
@@ -23,7 +22,7 @@ constexpr auto endian_flip_(std::size_t value, std::size_t n_relevant_bits) -> s
     return backward;
 }
 
-constexpr auto is_valid_marginal_bitstring_(const std::string& bitstring) -> bool
+auto is_valid_marginal_bitstring_(const std::string& bitstring) -> bool
 {
     const auto is_bitstring_char = [](char bitchar) {
         return bitchar == '0' || bitchar == '1' || bitchar == ket::internal::MARGINALIZED_QUBIT;
@@ -32,7 +31,7 @@ constexpr auto is_valid_marginal_bitstring_(const std::string& bitstring) -> boo
     return std::all_of(bitstring.begin(), bitstring.end(), is_bitstring_char);
 }
 
-constexpr auto is_valid_nonmarginal_bitstring_(const std::string& bitstring) -> bool
+auto is_valid_nonmarginal_bitstring_(const std::string& bitstring) -> bool
 {
     const auto is_nonmarginal_bitstring_char = [](char bitchar) {
         return bitchar == '0' || bitchar == '1';
@@ -41,18 +40,18 @@ constexpr auto is_valid_nonmarginal_bitstring_(const std::string& bitstring) -> 
     return std::all_of(bitstring.begin(), bitstring.end(), is_nonmarginal_bitstring_char);
 }
 
-constexpr void check_bitstring_is_valid_nonmarginal_(const std::string& bitstring)
+void check_bitstring_is_valid_nonmarginal_(const std::string& bitstring)
 {
     if (!is_valid_nonmarginal_bitstring_(bitstring)) {
         throw std::runtime_error {"Received bitstring with inputs that aren't '0' or '1'\n"};
     }
 }
 
-constexpr void check_bitstring_is_valid_marginal_(const std::string& bitstring)
+void check_bitstring_is_valid_marginal_(const std::string& bitstring)
 {
     if (!is_valid_marginal_bitstring_(bitstring)) {
         throw std::runtime_error {"Received bitstring with inputs that aren't '0', '1', or the marginal symbol.\n"};
     }
 }
 
-}  // namespace impl_ket
+}  // namespace ket::internal
